Use unsigned int (instead of short) for age
... short seems to translate to uint16, with maxvalue of 65k that's
dangerously low.

+++ src/utopia/models/SEIRD/SEIRD.hh
@@ -828,8 +828,7 @@ class SEIRD : public Model<SEIRD, CDTypes>
         _dset_age->write(_cm.cells().begin(),
                          _cm.cells().end(),
                          [](const auto& cell) {
-                             return static_cast<unsigned short int>(
-                                 cell->state.age);
+                             return cell->state.age;
                          });
 
         // Identify clusters and write them out
